+++ completion_aggregator/settings/common.py
@@ -11,23 +11,6 @@ def plugin_settings(settings):
     """
     Modify the provided settings object with settings specific to this plugin.
     """
-    # Emit feature allows to publish two kind of events progress and completion
-    # This setting controls which type of event will be published to change the default behavior
-    # the block type should be removed or added from the progress or completion list.
-    settings.ALLOWED_COMPLETION_AGGREGATOR_EVENT_TYPES = {
-        "progress": {
-            "course",
-            "chapter",
-            "sequential",
-            "vertical",
-        },
-        "completion": {
-            "course",
-            "chapter",
-            "sequential",
-            "vertical",
-        }
-    }
     settings.COMPLETION_AGGREGATOR_BLOCK_TYPES = {
         'course',
         'chapter',
@@ -35,6 +18,12 @@ def plugin_settings(settings):
         'vertical',
     }
 
+    # Emit feature publishes progress events to track aggregated completion.
+    # Defaults to the full set of block types subject to completion aggregation.
+    # Block types may be removed from this list to limit the tracking log events emitted.
+    settings.COMPLETION_AGGREGATOR_TRACKING_EVENT_TYPES = settings.COMPLETION_AGGREGATOR_BLOCK_TYPES
+
+    # Synchronous completion aggregation is enabled by default
     settings.COMPLETION_AGGREGATOR_ASYNC_AGGREGATION = False
 
     # Names of the batch operations locks
@@ -58,10 +47,9 @@ def plugin_settings(settings):
 
     # Whitelist the aggregator events for use with event routing backends xAPI backend.
     enabled_aggregator_events = [
-        f'openedx.completion_aggregator.{event_type}.{block_type}'
+        f'openedx.completion_aggregator.progress.{block_type}'
 
-        for event_type in settings.ALLOWED_COMPLETION_AGGREGATOR_EVENT_TYPES
-        for block_type in settings.ALLOWED_COMPLETION_AGGREGATOR_EVENT_TYPES[event_type]
+        for block_type in settings.COMPLETION_AGGREGATOR_TRACKING_EVENT_TYPES
     ]
     settings.EVENT_TRACKING_BACKENDS_ALLOWED_XAPI_EVENTS += enabled_aggregator_events
     settings.EVENT_TRACKING_BACKENDS.update(event_tracking_backends_config(